refactor: 관심 주식 페이지 헤더 구조 변경
